split code into 3 modules
